Started to implement new concept for data threading. Still need to work on analog output and eventdata.
git-svn-id: https://svn.code.sf.net/p/relacs/code/relacs/trunk@1043 1550813e-0d35-0410-aec7-fc86b090456a

+++ plugins/efield/src/eodtools.cc
@@ -36,7 +36,7 @@ double EODTools::eodThreshold( const InData &data, double tbegin, double tend,
   contrast = fabs( contrast );
   if ( contrast > 0.99 )
     contrast = 0.99;
-  double min=0.0, max=0.0;
+  float min=0.0, max=0.0;
   data.minMax( min, max, tbegin, tend );
   return 0.75*(max-min)*(1.0-contrast)/(1.0+contrast);
 }
